changes to sac, works on terminating envs now
Former-commit-id: 23ff3f58cf408a9e8c4968a6eb6207a7329b5c6d
Former-commit-id: 3e482da34152e062dc0ba6dbc158556e036288d0 [formerly d2ae32a576f8d864a3d8713ded19fc4db58d1813] [formerly b49f5493471c4d81fee66e4276c83cd19adb273c [formerly 9705bf0187b100ca5594c1f9e91d1648c9d8451a]]
Former-commit-id: eb7294dd26e70584186ea366e8d1cf68922e872d [formerly c48f33be296a21c84e8086b747bc49804ab0d8ed]
Former-commit-id: 384a4a9ea018ca92959c54ee84e42a37c7484b3d

+++ seagul/envs/classic_control/sg_cartpole.py
@@ -57,11 +57,11 @@ class SGCartPoleEnv(gym.Env):
         self.init_state_noise_max = 0.1
         high = np.array([2 * pi, self.X_MAX, self.DTHETA_MAX, self.DX_MAX])
         low = -high
-        self.observation_space = gym.spaces.Box(low=low, high=high)
+        self.observation_space = gym.spaces.Box(low=low, high=high,dtype=np.float32)
 
         self.TORQUE_MAX = 5.0
         self.torque_noise_max = 0.0
-        self.action_space = gym.spaces.Box(-self.TORQUE_MAX, self.TORQUE_MAX, shape=(1,))
+        self.action_space = gym.spaces.Box(-self.TORQUE_MAX, self.TORQUE_MAX, shape=(1,), dtype=np.float32)
 
         self.viewer = None
